requirements.txt
discord.py>=2.4.1
aiosqlite>=0.18.0
python-dotenv>=1.0.0